requirements.txt
altgraph==0.17.4
charset-normalizer==3.4.2
colorama==0.4.6
packaging==24.2
pefile==2023.2.7
pillow==11.3.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
python-dotenv==1.1.0
pywin32-ctypes==0.2.3
reportlab==4.4.2
setuptools==78.1.0
tabulate==0.9.0
